moved tests directory out of src
